Fixes the compilation for Cortex-M example
The addition of printf() broke the compilation for cortex-m55 example.
This is because printing long long int requires doing mod with 64-bit
integers. There is no assembly instruction to do so and needs to be
supplied with assembly subroutines which handle those operations.

libgcc supplies with such subroutines for printf to work correctly.

Signed-off-by: Piyush Itankar <itankar@google.com>