Allows dynamic setting of input/output streams
user needs to provide pointers to functions to be used for getting
a byte from a source and sending a byte to the destination.

This would allow for easy implementation of mixed use cases like:
read from uart but write to memory. Or read from one source and
write to many.

Signed-off-by: Piyush Itankar <itankar@google.com>
Change-Id: I1e61081156a96d510f04faecb54da228398e3bab