Allows disabling the command history feature
By default the command history feature is enabled. But users would
have the option to disable the feature if they do not need it.

Few controllers have less than 4kb of memory and it becomes
important to leave out features that can be optional.

The feature can be disabled by adding the following in project level
makefile -

DEFINES += -DSHELL_NO_HISTORY

Signed-off-by: Piyush Itankar <itankar@google.com>
Change-Id: I5d5b829e591481012721e38294886856154827dc