Adds support for specifying folders to search for C files
This can be useful when we want to use something that is not inside the
project folder - for example, shared drivers/tests between a different
project and the early-bringup based project.

Also adding support to ignore certain files. This will help as currently
all files are compiled, and rules.mk (or such files) are ignored.

Change-Id: I32fed76bd185889ffdd6c489faa48a56e44d13b3
Signed-off-by: Aneesh D H <aneeshdh@google.com>