Spinlock library support.
Spinlock supports guarding of the critical section in multi-core system.

Adds a spinlock library which supports initialization of spinlock,acquiring of spinlock, release of spinlock and supports freezing of cores in case of re-entrancy.

Adds a multicore_printf to provide a spinlock wrapper around printf() to avoid interleaving concurrent printf's.

Change-Id: Id206ca37af73446311093e77e937dcca260085a4