Spinlock Optimization: Ticket Spinlock
In ticket spinlock the waiting cores are processed in a first-in first-out(FIFO) basis using ticket-based algorithm.
As a result it increases fairness and prevents starvation.

Adds a next ticket and current ticket variable to spinlock structure.
Adds a lock status variable to spinlock structure.
Adds a GetTicket and CheckTicket loop in acquire of spinlock.
Adds a ticket incrementing mechanism in release of spinlock.

Change-Id: Ia5e0dab8d36eb337c5f847f3e4c6f1625b3eae1d