Use LOADADDR and ADDR for defining symbols in .ld
This is more robust and should always work.
The previous way of defining symbols can fail when there are orphan
sections. An example would be when there is a section which has exec
properties - the linker may choose to place this between .text and .data
and this will mean data is not initialised properly.

Change-Id: Ib7199e0b83b81f4660fab159bf6d8802e0f619c4
Signed-off-by: Aneesh D H <aneeshdh@google.com>